+++ addOns/openapi/src/test/java/org/zaproxy/zap/extension/openapi/automation/OpenApiJobUnitTest.java
@@ -30,6 +30,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.withSettings;
 
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
@@ -90,11 +91,12 @@ class OpenApiJobUnitTest extends TestUtils {
         Map<String, String> params = job.getCustomConfigParameters();
 
         // Then
-        assertThat(params.size(), is(equalTo(4)));
+        assertThat(params.size(), is(equalTo(5)));
         assertThat(params.get("apiFile"), is(equalTo("")));
         assertThat(params.get("apiUrl"), is(equalTo("")));
         assertThat(params.get("targetUrl"), is(equalTo("")));
         assertThat(params.get("context"), is(equalTo("")));
+        assertThat(params.get("user"), is(equalTo("")));
     }
 
     @Test
@@ -105,6 +107,7 @@ class OpenApiJobUnitTest extends TestUtils {
         String apiUrl = "https://example.com/test%20file.json";
         String targetUrl = "https://example.com/endpoint/";
         String context = "My Context";
+        String user = "My User";
         String yamlStr =
                 "parameters:\n"
                         + "  apiUrl: "
@@ -117,11 +120,18 @@ class OpenApiJobUnitTest extends TestUtils {
                         + targetUrl
                         + "\n"
                         + "  context: "
-                        + context;
+                        + context
+                        + "\n"
+                        + "  user: "
+                        + user;
         Yaml yaml = new Yaml();
         Object data = yaml.load(yamlStr);
 
+        AutomationEnvironment env = mock(AutomationEnvironment.class);
+        given(env.getAllUserNames()).willReturn(List.of(user));
+
         OpenApiJob job = new OpenApiJob();
+        job.setEnv(env);
         job.setJobData(((LinkedHashMap<?, ?>) data));
 
         // When
@@ -133,6 +143,7 @@ class OpenApiJobUnitTest extends TestUtils {
         assertThat(job.getParameters().getApiUrl(), is(equalTo(apiUrl)));
         assertThat(job.getParameters().getTargetUrl(), is(equalTo(targetUrl)));
         assertThat(job.getParameters().getContext(), is(equalTo(context)));
+        assertThat(job.getParameters().getUser(), is(equalTo(user)));
         assertThat(progress.hasErrors(), is(equalTo(false)));
         assertThat(progress.hasWarnings(), is(equalTo(false)));
     }